Moving properties to common project
