+++ api/src/main/kotlin/com/example/todo/domain/service/TodoAuthenticationUserDetailService.kt
@@ -1,7 +1,7 @@
 package com.example.todo.domain.service
 
-import com.example.todo.domain.configuration.TodoAppApiKey
 import com.example.todo.enums.UserRole
+import com.example.todo.properties.TodoAppApiKey
 import org.springframework.security.core.authority.AuthorityUtils
 import org.springframework.security.core.userdetails.AuthenticationUserDetailsService
 import org.springframework.security.core.userdetails.User
