+++ web/src/main/resources/application.yml
@@ -1,3 +1,6 @@
 spring:
   profiles:
     active: local
+    group:
+      local:
+        - common
